+++ src/main/java/nextstep/subway/ui/ControllerExceptionHandler.java
@@ -21,19 +21,3 @@ public class ControllerExceptionHandler {
         return ResponseEntity.unprocessableEntity().body(errorMessage);
     }
 }
-
-class ErrorMessage {
-    private String message;
-
-    static ErrorMessage of(String message) {
-        return new ErrorMessage(message);
-    }
-
-    public ErrorMessage(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
